Do not accept `none` for `iconType` prop
It makes no sense, because it will basically turn off the legend.
Including `none` as a valid option would be a bit deceiving, making it
seem like just the icon will not be rendered, but instead, the whole
legend will turn off.  It only makes sense to set `none` on particular
items, not the whole legend.