Use item's data before using chart's data.
Changed the way `getDisplayedData` finds the data to be used. It will
try to return the item's data before using the chart's data.

I did this change because I believe the chart's data is less important
the a item's data, when it is present, and this allows me to use a
composed chart with the Line component getting its data from the chart
and the Scatter component getting its data from its own properties.