Fix tooltips that disappear while mouse still over a scatter point
Reported here: https://github.com/recharts/recharts.org/issues/45

Reproducible in this jsfiddle:

https://jsfiddle.net/alidingling/uLysj0u2/

This happens because the cursor draws on top of the scatter. Once it appears, it starts snagging mouse events which belong to the scatter. The behavior becomes even more pronounced as the scatter dots get smaller.

It probably does not make sense for the cursor ever to receive pointer events.