Allow to add more classes in tooltips
Since #798 was fixed, there was no way to change the tooltip wrapper
style other than defining another class in css for
"recharts-default-tooltip" and mark everything as "important.

The problem with that approach is that we would be changing all the
tooltips, but what if we want to change it for only one chart?

This will allow to define extra classes  for both the wrapper and the
label, allowing us to customize a single chart easily.