feat: Short-circuit axis map creation for specified domain
This is a hack to short-circuit the domain creation here to enhance performance. Usually, the data is used to determine the domain, but when the user specifies a domain upfront (via props), there is no need to calculate the domain start and end, which is very expensive for a larger amount of data.
The only thing that would prohibit short-circuiting is when the user doesn't allow data overflow, because the axis is supposed to ignore the specified domain that way.