Eslint react hooks plugin (#3466)
## Description

React hooks eslint plugin was installed but wasn't used. From now on,
it'll be extended from in the eslint config, so migrating components to
funtional components and usage of hook will be smoother as it can detect
misuse of hooks.

## Related Issue

## Motivation and Context

React hooks has some rules, when followed they are great utilities to
re-use code between function components and they're very powerful. But
when these ruled are ignored, they can cause regression, unnecessary
computation, or give unexpected results. `eslint-plugin-react-hooks` is
the eslint plugin to enforce these rules and make sure the code is
correct.

## How Has This Been Tested?

## Screenshots (if appropriate)

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.