[a11y] Keyboard navigation support for categorical charts (#3546)
<!--- Provide a general summary of your changes in the Title above -->

## Description

Add an `accessibilityLayer` prop to categorical charts. This layer will
modify the SVG element to add the chart to the tab order. Then, when
users focus on the chart and press arrow keys, the accessibility layer
spoofs mouse movements to force the tooltips and active nodes to appear
along the chart.

This process, where the accessibility layer effectively helps a
keyboard-only user pretend to be a mouse user, was the least-change
solution that still enabled keyboard access to the contents and
interactions of the chart.

Test cases were updated, with 100% code coverage on new files that were
added.

Documentation has also been updated, which explains to developers what
the purpose of the accessibilityLayer feature is, and how to use it
effectively.

<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: Julianna Langston <julianna@cosaic.io>