add: initial width & height to ResponsiveContainer
ResponsiveContainer component is used to render charts that can adapt to
their container. When a chart is rendered inside Responsive container,
you can omit its own height and width, and instead specify the
dimensions on the Container, one of the dimensions should be a
percentage, using `%` symbol, for instance 50%, recharts then tries to
to render the chart accordingly. This is possible because when the
component renders in the browser, then it'll calculate the container's
dimensions, and then it'll calculate the width (or the height) that was
specified in percentage.

When this ResponsiveContainer renders initially, its internal sizes are
sepcified as `{ width: -1, height: -1 }`, this is because it needs to
render in the browser before it can calculate the dimentions. Also,
ResponsiveContainer doesn't render the chart initialy since the
dimentions are negative, so it'll try to find a positive height and
width and render the chart accordingly.

This is important, because it means that, ResponsiveContainer relies on
JavaScript and Browser functionalities to calculate its children size,
and initially, it doesn't render the children. This can be a limitation
when the ResponsiveContainer functionality is needed but the
charts are rendered in an environment which is not the browser
or JavaScript is not available.

In this commit, I'm changed this behavior such that, initial height and
initial width can be specified to calculate the initial dimentions of
the ResponsiveContainer's child. These props, if not provided, they
have the original sizes (-1, -1) respectively, so this changes can be
applied without breaking semver.

Which means, its possible to render the charts in specified dimention
while rendering on server, and inital render on browser (before
hydration), and when hydrated, the sizes will be calculated as
before.