migrate generateCategoricalChart to functional
This is an attempt to migrate the component inside
generateCategoricalChart to a functional component.
I cloned the file and renamed it so I can get one of the
charts to work as expected as the current implementation

NOTE: Currently only LineChart uses the new implementation
I use it as a reference, later on we can only switch the function
call and that will be it, since the input and output is identical
to the current one, the difference is in the implementation
detail.