Address issue where tooltips and AccessibilityManager can get misaligned (#3627)
Addresses bug with AccessibilityManager where, when the user interacts
with the AccessibilityManager and *then* the `<Tooltip/>` component
changes, there is a javascript runtime error. (Note that I could only
reproduce this if the container chart used the `margin` prop.)

## Description

- I added error handling to prevent the javascript runtime error from
throwing.
- TDD test

## Related Issue

@ckifer - Was the issue filed? 

For reference, the issue can be reproduced using this sandbox:

https://codesandbox.io/s/stacked-area-chart-forked-8ipx36?file=/src/App.tsx:951-1010

## Motivation and Context

Solves a javascript runtime error

## How Has This Been Tested?

I generated a reproducible example within the jest testing framework and
in Storybook. Once I added in defensive checks for the error condition,
I confirmed that this solved the error in both a browser and jest.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [NA] My change requires a change to the documentation.
- [NA] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

Co-authored-by: Julianna Langston <julianna@cosaic.io>