fix: dont throw when dot is null, use typeguard in line and improve in area (#3637)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

When the `clipDots` PR was merged I added a typeguard to `Area` but I
did not see to do the same for `Line`. If `dot` was something other than
`DotProps` this cast is incorrect and wrong.

Linked issue is caused when passing `null` to `dot`. `null` is not an
acceptable type to `dot` and therefore this wasn't a breaking change,
but it still uncovered a need for this check.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

fixes https://github.com/recharts/recharts/issues/3635

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

See issue, removes need to cast to `DotProps`

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

Added unit tests that reproduce the issue and are green after fixing it

## Screenshots (if appropriate):

N/A no visual changes

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [Y] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [Y] My code follows the code style of this project.
- [N] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [Y] I have added tests to cover my changes.
- [Y] All new and existing tests passed.

Co-authored-by: Coltin Kifer <ckifer@amazon.com>