Switches useLayoutEffect to useEffect
When rendering a chart on server side we get this warning:

Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
    at Rectangle (app/node_modules/recharts/lib/shape/Rectangle.js:81:35)

Switching to useEffect to solve the problem and the chart renders well without any warnings.