Storybook and Type Information: Move description next to type and reuse existing (#3673)
In the ideal setup, storybook use [automatic arg types
inference](https://storybook.js.org/docs/react/api/arg-types#automatic-argtype-inference)
via
[react-docgen-typescript](https://github.com/styleguidist/react-docgen-typescript).

We instead have mostly duplicated documentation by creating separate
props documentation in the storybook folder, such as AnimationDuration.

Instead, we should define our own types, and add a comment right above
to describe/explain. To do so, requires us to define a type once, and
reuse it elsewhere.

For `animationDuration: number`, to not redundantly redefine the
description per comment, I define an alias `type AnimationDuration =
number` with a description above. Using this type everywhere, leads to
the description being added to every component in the ArgsTable in
storybook.

This PR is a suggestion how to move forward and move documentation from
storybook into the codebase.

This PR only tackles two properties: 
- AnimationDuration
- AnimationTiming

If we agree this is a reasonable way forward, I will follow up with
other properties.

<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.