add: default tabIndex and accept it as prop
Previously the root layer of Pie was always keyboard focus-able which
means it wasn't possible to control its behavior. This changes that to
accept an optional tabIndex prop which by default is `0` the old value
to make sure not introduce a breaking change, except now users can
control this via a prop.