Bugfix/b(#2778) tooltip position when container uses transform scale (#3748)
Fixes detection of a tooltip mouse event

## Description

When a parent container used transform: scale(Number) Css - the tooltip
mouse event was fired up in the incorrect mouse position

## Related Issue

https://github.com/recharts/recharts/issues/2778

## Motivation and Context

This changes allows user to style parent container with transform style
css property without affecting tooltip behaviour

## How Has This Been Tested?

npm run test, npm run storybook and did some manual testing. I also
created a dedicated story.
The changes should not affect other parts of the code - it was a change
within 1 function, that already had tests around it. It introduces new
parameter to a "inRange" function that have default value if wasn't
passed. The API of that function was not changed, only extended.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] I have added a story to the storybook
- [x] All new and existing tests passed.

---------

Co-authored-by: Mateusz Sidorski <mateusz.sidorski@cardinality.co.uk>