feat: add onDragEnd event to Brush component (#3774)
<!--- Provide a general summary of your changes in the Title above -->

Currently the only way to access event data from `Brush` is through
`onChange` event handler which is fired whenever the brush is changed.
This is not ideal as it fires whenever the brush moves and when the
chart needs to load data from an API for an example this might become
slow. As mentioned in linked issues, workaround might be to debounce
whatever is needed to be done after dragging the brush is done but it is
still just a workaround.

## Description

<!--- Describe your changes in detail -->
This change is adding `onDragEnd` function to the `Brush` component and
exposed start and end indexes through it.

## Related Issue

https://github.com/recharts/recharts/issues/2534
https://github.com/recharts/recharts/issues/859

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
I have created a new storybook example that utilises the existing Brush
and newly added onDragEvent. In the example i've added text component
that keeps track of currently active brush indexes. After the brush has
been updated and dragging has ended the indexes updated to the latest
ones.

## Screenshots (if appropriate):
N/A

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] I have added a storybook story or extended an existing story to
show my changes
- [x] All new and existing tests passed.

---------

Co-authored-by: Nikolas Rieble <nikolas-rieble@protonmail.com>