CartesianAxis: Improve interval option 'equidistantPreserveStart' (#3768)
<!--- Provide a general summary of your changes in the Title above -->

## Description

This PR overhauls the existing interval option
`equidistantPreserveStart`

In this variant, we always show the first tick, and starting from the
first, we find the smallest N, for which all n-th ticks can be shown
without collision.

A collision is defined as two overlapping ticks, taking into account
custom tick formatting, a unit, the fontsize, the angle, et cetera (in
alignment with the other existing methods). An overlap is identified if
the end of the previous tick overlaps with the start of the next one.

Downside of this method: 
- When the first tick is very long, hardly any other ticks will be
shown.
- Performance: Instead of iterating over the ticks once, as all other
methods do, this method iterates at most M times, M being the total
amount of available ticks (i.e. for 1000 datapoints, 1000 times). All
other methods are thus more performant. This is a good candidate to
discuss in review or to iterate upon.

<!--- Describe your changes in detail -->

## Related Issue
https://github.com/recharts/recharts/issues/3305

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [ ] All new and existing tests passed.

---------

Co-authored-by: “Nikolas <“nikolas@rieble.com“>