Move getCursorPoints and getCursorRectangle to their own files (#3851)
And add unit tests

## Description

## Related Issue

https://github.com/recharts/recharts/issues/3823

https://github.com/recharts/recharts/discussions/3717

## Motivation and Context

For the PoC I want to call these two functions from outside of
`getCategoricalChart` but I figured that master will benefit from this
change too: 1. it makes the mammoth file a bit smaller and 2. makes it
easier to unit test so I did that too.

Also got rid of unnecessary lodash call in the spirit of
https://github.com/recharts/recharts/issues/3752

## How Has This Been Tested?

Jest, TS

## Screenshots (if appropriate):

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [X] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [X] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [X] All new and existing tests passed.