Remove unit test that was comparing new/old reduce-css-calc
Why? It's not testing anything - it throws, and the error is not reported as a failing test but instead as a log message in the console.

Instead I have replaced it with a `.not.toThrow()` call which at least tests that our new implementation is correct.
The old one throws so I removed it completely.