Reduce calls to `getStringSize` when calculating visible ticks (#2589) (#3953)
<!--- Provide a general summary of your changes in the Title above -->

## Description

Reduce the number of times `getStringSize` is called when calculating
which ticks are rendered.
This optimization is applied to `getTicksEnd, `getTicksStart` and
`getEquidistantTicks`, but the latter is unlikely to benefit from it
since the algorithm does not iterate over the candidate ticks
sequentially.

I have created [this
repository](https://github.com/bernardobelchior/recharts-repro), based
on the JSFiddle available in [this
issue](https://github.com/recharts/recharts/issues/1465#issue-356913907)
to test the changes. It generates a chart with 10000 random data points.

Based on Recharts' master branch (pointing to commit
`8cdfab7facc4d1d107c11a1fc3d4a26e716dddca` at the time of writing), the
Chrome DevTools' Performance chart is as follows:


![image](https://github.com/recharts/recharts/assets/12778398/52da2b7b-0cef-4ed4-9d47-a9b9fc2d23f2)


You can see that 95% time is taken by `getStringSize()` and descendants,
which account for 3.2 seconds on my M1 Macbook Pro from 2020. In fact, I
counted and `getStringSize()` is being called around 40k times to render
the chart.


Since we know that `getStringSize` is expensive, whatever we can do to
reduce the number of times it's called will likely yield a performance
improvement. As such, this PR is comprised of two changes:
1. Call `getTickSize` in `getTicksEnd`, `getTicksStart` and
`getEquidistantTicks` only when absolutely necessary;
1. Update the `isVisible` calculation to check if the tick won't be
visible even before we take size into account: this way we can skip
calculating the string size if we already know that the tick won't be
visible.


After this change, `getStringSize` still takes up most of the time
(88%), but the absolute value is now much smaller (1.8 seconds):


![image](https://github.com/recharts/recharts/assets/12778398/10dd9067-a89f-4b95-b2d2-e22887345dde)

`getStringSize` is now called around 18k times, which means there's
still a lot of room for future improvements.


## Related Issue

#2589 
#1465
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

Explained above

## How Has This Been Tested?

Added some unit tests. There were none, so it's hard to see if there's
no regression, but I ran the new tests against the `isVisible` function
before and after these changes and the results were the same.

If there's any edge case you'd like me to test, please let me know 😄 
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->


## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [x] All new and existing tests passed.