Custom Storybook theme for Recharts brand (#4016)
## Description

Creates a custom Storybook theme using the colors, fonts, and logo from
Recharts.org. Links to Recharts.org if user clicks on the logo.

Note: I grabbed a screenshot of the logo from the website. Might want to
replace with a better image if one is available.

## Related Issue

Closes #3446. 

## Motivation and Context

Customizes Storybook to better line up with the website branding. 

## How Has This Been Tested?

This only affects the Storybook theme, not the function of any chart
components. All tests are passing.


## Types of changes

- [ X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [X ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [ X] All new and existing tests passed.